Prosemirror HTML handling
This implements handling of images in pasted HTML when running with
Prosemirror.

When a user pastes HTML which contains images, the plugin will try to
download those images and upload them as if they had been pasted
directly. For this to work, the images need to be references as full
URLs in the HTML and need to be publically accessible (eg. the DokuWiki
HTTPClient needs to be able to download them).

This feature has been mostly tested with copy'n'paste from Google Docs
but should also work when copy'n'pasting from arbitrary websites.